moved Data table to quiz app
- wanted to have everything under /quiz
