+++ lifeapi_apps/base_app/urls.py
@@ -3,5 +3,4 @@ from . import views
 
 urlpatterns = [
     path("", views.home, name="home"),
-    path("data_table/", views.data_table, name="data_table")
 ]
